refactor: Pull the serac/physics/opererator files into serac/numerics
