+++ src/serac/numerics/functional/tests/functional_comparisons_cuda.cu
@@ -15,7 +15,7 @@
 #include "serac/infrastructure/input.hpp"
 #include "serac/serac_config.hpp"
 #include "serac/numerics/expr_template_ops.hpp"
-#include "serac/physics/operators/stdfunction_operator.hpp"
+#include "serac/numerics/stdfunction_operator.hpp"
 #include "serac/numerics/functional/functional.hpp"
 #include "serac/numerics/functional/tensor.hpp"
 #include "serac/infrastructure/profiling.hpp"
