+++ src/serac/numerics/stdfunction_operator.hpp
