+++ src/serac/numerics/tests/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause) 
 
-set(test_dependencies gtest serac_numerics)
+set(test_dependencies gtest serac_numerics serac_boundary_conditions serac_state)
 blt_list_append( TO test_dependencies ELEMENTS caliper IF ${SERAC_USE_CALIPER} )
 
 set(numerics_parallel_tests
@@ -18,6 +18,8 @@ serac_add_tests( SOURCES ${numerics_parallel_tests}
 set(numerics_serial_tests
     expr_templates.cpp
     mesh_generation.cpp
+    serac_operator.cpp
+    serac_odes.cpp
     )
 
 serac_add_tests( SOURCES ${numerics_serial_tests}
