+++ src/serac/numerics/tests/serac_odes.cpp
@@ -13,8 +13,8 @@
 #include "mfem.hpp"
 
 #include "serac/numerics/expr_template_ops.hpp"
-#include "serac/physics/operators/odes.hpp"
-#include "serac/physics/operators/stdfunction_operator.hpp"
+#include "serac/numerics/odes.hpp"
+#include "serac/numerics/stdfunction_operator.hpp"
 
 using namespace serac;
 using namespace serac::mfem_ext;
