+++ src/serac/physics/CMakeLists.txt
@@ -4,7 +4,6 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 
-add_subdirectory(operators)
 add_subdirectory(integrators)
 add_subdirectory(materials)
 add_subdirectory(state)
@@ -34,7 +33,6 @@ set(physics_dependencies
     serac_physics_integrators
     serac_physics_materials
     serac_numerics
-    serac_physics_operators
     serac_boundary_conditions
     serac_state
     )
