+++ src/serac/physics/solid.hpp
@@ -18,8 +18,8 @@
 
 #include "serac/infrastructure/input.hpp"
 #include "serac/physics/base_physics.hpp"
-#include "serac/physics/operators/odes.hpp"
-#include "serac/physics/operators/stdfunction_operator.hpp"
+#include "serac/numerics/odes.hpp"
+#include "serac/numerics/stdfunction_operator.hpp"
 #include "serac/physics/materials/hyperelastic_material.hpp"
 #include "serac/physics/integrators/displacement_hyperelastic_integrator.hpp"
 #include "serac/coefficients/sensitivity_coefficients.hpp"
